dlS-9792 update directory structure
